back-end/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
cinemagoer==2022.12.27
    # via -r requirements.in
click==8.1.3
    # via flask
flask==2.2.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==3.0.10
    # via -r requirements.in
importlib-metadata==6.1.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
lxml==4.9.2
    # via cinemagoer
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector==2.2.9
    # via -r requirements.in
six==1.16.0
    # via flask-cors
sqlalchemy==2.0.7
    # via cinemagoer
typing-extensions==4.5.0
    # via sqlalchemy
werkzeug==2.2.3
    # via flask
zipp==3.15.0
    # via importlib-metadata
